[refactor](execution) Extract shared LocalEngineExecutor base
Move the engine-agnostic executor machinery (background event loop, task
counting, submit/take/wait/shutdown, per-executor routing, abort/release)
out of LocalVLLMExecutor into a shared LocalEngineExecutor base. vLLM and
SGLang executors now supply only four backend hooks (_engine_name,
_materialize_sampling_params, _create_engine, _run_generate), eliminating
SGLang's copy of the machinery.
